requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asarPy==1.0.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Cython==3.0.12
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-genai==1.10.0
google-generativeai==0.8.4
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Mako==1.3.9.dev0
Markdown==3.5.2
MarkupSafe==3.0.2
meson==1.5.2
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.27.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyGObject==3.50.0
pyparsing==3.2.3
requests==2.32.3
rsa==4.9
setuptools==78.1.0.post20250409
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
websockets==15.0.1
wheel==0.46.0
